+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#469](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/469))
 - Implemented specification that HTTP span attributes must not contain username and password.
   ([#538](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/538))
+- Changed the psycopg2-binary to psycopg2 as dependency in production
+  ([#543](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/543))
 
 ### Added
 - `opentelemetry-instrumentation-httpx` Add `httpx` instrumentation